requirements.txt
Django==3.1
djangorestframework==3.11.1
python-dotenv==0.14.0

